ref(data,server,handlers): remove the Count interface
Makes progress toward
https://github.com/deis/workflow-manager-api/issues/61
