+++ get_latest_versions_test.go
@@ -20,7 +20,7 @@ func TestGetLatestVersions(t *testing.T) {
 	db, err := data.VerifyPersistentStorage(memDB)
 	assert.NoErr(t, err)
 	versionFromDB := data.VersionFromDB{}
-	srv := newServer(memDB, versionFromDB, data.ClusterCount{}, data.ClusterFromDB{})
+	srv := newServer(memDB, versionFromDB, data.ClusterFromDB{})
 	defer srv.Close()
 	const numComponentVersions = 6
 	components := make(map[string]types.ComponentVersion)
